fix(model): force LiDAR backbone to run in fp32 (P2 fp16 mixed precision compat)
P2 smoke test 报错:
  RuntimeError: Input type (torch.cuda.HalfTensor) and weight type
  (torch.cuda.FloatTensor) should be the same  (in SECOND.forward conv2d)

根因:Sparse4D config 里 fp16=dict(loss_scale=32.0),image backbone 全程 fp16,
但 LiDAR backbone (SparseEncoder + SECOND + SECONDFPN) 的 weights 是 fp32。
data['points'] 经 fp16 hook 转成 half,进入 voxelize → SparseEncoder 后,
SECOND conv2d 收到 fp16 input vs fp32 weight,type mismatch 崩。

修复:
- voxelize 加 @force_fp32(),输入 cast 成 fp32 处理
- extract_lidar_feat 入口 [p.float() for p in points] 强制 fp32
- DFA _sample_lidar_bev 之后,lidar_features.to(features.dtype) cast 回
  image features 的 dtype(fp16),保证 sum / MAFS fusion 不 type mismatch

这是 mmdet3d / CenterPoint / BEVFusion 等 SOTA LiDAR 工作的标准做法 —
LiDAR backbone 用 fp32 (spconv fp16 支持不完整),image backbone 用 fp16,
两路在 fusion 层重新 dtype align。

+++ projects/mmdet3d_plugin/models/blocks.py
@@ -200,9 +200,12 @@ class DeformableFeatureAggregation(BaseModule):
         # ============ Phase 2 F1/F2: LiDAR BEV fusion ============
         # 在 output_proj 之前 fuse,这样 image / lidar features 在同一抽象层
         if self.lidar_bev_sampling and metas.get("lidar_bev") is not None:
-            lidar_bev = metas["lidar_bev"]  # (B, C, H, W)
+            lidar_bev = metas["lidar_bev"]  # (B, C, H, W),来自 fp32 LiDAR backbone
             lidar_features = self._sample_lidar_bev(anchor, lidar_bev)
             lidar_features = self.lidar_proj(lidar_features)  # (B, A, embed_dims)
+            # Cast 回 features 的 dtype(在 fp16 训练下 features 是 half,
+            # lidar_features 默认是 fp32,直接相加会 type mismatch)
+            lidar_features = lidar_features.to(features.dtype)
 
             if self._fusion_mode == "sum":
                 # F1: simple sum (固定 50:50 fusion)
